@@ -9,25 +9,29 @@
     "case",
     "renamer-plugin"
   ],
+  "main": "./index.mjs",
+  "exports": "./index.mjs",
   "licence": "MIT",
   "engines": {
-    "node": ">=8"
+    "node": ">=14"
   },
   "scripts": {
-    "test": "test-runner test.js",
-    "cover": "nyc test-runner test.js && nyc report --reporter=text-lcov | coveralls"
+    "test": "test-runner test.mjs",
+    "cover": "c8 npm test && c8 report --reporter=text-lcov | coveralls"
   },
   "files": [
-    "index.js"
+    "index.mjs"
   ],
   "dependencies": {
-    "lodash": "^4.17.10"
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "coveralls": "^3.0.2",
+    "c8": "^7.7.1",
+    "coveralls": "^3.1.0",
+    "current-module-paths": "^0.1.0",
     "mkdirp2": "^1.0.4",
-    "renamer": "^1.0.0-5",
-    "rimraf": "^2.6.2",
-    "test-runner": "^0.5.0"
+    "renamer": "^3.0.0-6",
+    "rimraf": "^3.0.2",
+    "test-runner": "^0.9.2"
   }
 }
